Create local variable.c
local variables are variables  that are declared inside a function or block are called local variables. They can be used only by statements that are inside that function or block of code. Local variables are not known to functions outside their own. The above example shows how local variables are used. above all the variables a, b, and c are local to main() function.